decNumber CFLAGS change from -Wall to -Wall -Wextra -Wno-maybe-initialized
This is a workaround for a compiler bug in recent GCC.  It spent a lot
of time looking at this every which way and determined the compiler
must be wrong.  Looking at GCC BugZilla, it appears that others have
also analyzed this and come to the same conclusion - see GCC BZ#95681.

This quiets the false positive warning and is tested on GCC and Clang.

Upstream-bug-URL: https://gcc.gnu.org/bugzilla/show_bug?id=95681