Update GNUmakefile - when using gcc, add -Wno-stringop-overread to CFLAGS
This is another bug in GCC, tracked upstream as GCC BZ#90132, but also a
hairy bit of code in decNumber I don't want to change and test to
silence, so, just squelch the warning when using GCC.  The other
workaround would be to use -O2 but the performance hit is large.

GCC is closed the bug WONTFIX - it seems for this bug their own GCC
bootstrap just silences the warning to appease Werror now.

Also, upstream GCC is also considering dropping decNumber completely,
which means it could become a problem soon. (GNU and IBM both say)
"decNumber is barely maintained and is quite a big mess."

Previously, I had brought up MKL/MKL2/IPL, but there was little to be
gained and a lot of regression testing needed for no real benefit. This
might be revisited with automated regression testing now available.

There were also license concerns which are now allevianced with the
license change in mid-2017 (ISSL) which allows for redistribution and is
compatible with the ICU, MIT, BSD, X11, and similar variant licenses,
but is not compatible with the GPL (which is fine for our case, but
should be noted).

This change would be something out in the future but would be pressing
if IBM official drops decNumber or it would not be used for GCC
bootstrap at IBM as it would quickly cause issues and requires a deep
understanding of the IEEE spcs that I don't have (or have time for),
and likely nobody else else.  For now, thought, decNumber is still the
best game in other than MKL/IPL.

Disassembly and analysis at \
       https://ggcc.gnu.org/bugzilla/show_bug.cgi?id=90132